+++ lib/td_dd/lineage/graph_data/nodes.ex
@@ -3,13 +3,16 @@ defmodule TdDd.Lineage.GraphData.Nodes do
   Functions for handling graph data node queries.
   """
   alias Graph.Traversal
+  alias TdDd.DataStructures.DataStructure
   alias TdDd.Lineage.GraphData.State
+  alias TdDd.Lineage.Units
+  alias TdDd.Lineage.Units.Node
 
   @doc """
   Returns the context of a given external id within hierarchy graph. The context
   consists of a list of parent nodes with the siblings of the selected child.
   """
-  def query_nodes(external_id, %State{contains: t, roots: roots}) do
+  def query_nodes(external_id, user, %State{contains: t, roots: roots}) do
     case path(t, external_id) do
       :not_found ->
         {:error, :not_found}
@@ -17,7 +20,7 @@ defmodule TdDd.Lineage.GraphData.Nodes do
       path ->
         res =
           [nil | path]
-          |> Enum.map(&{&1, children(t, &1, roots)})
+          |> Enum.map(&{&1, children(t, &1, roots, user)})
           |> Enum.map(fn {parent, m} -> Map.put(m, :parent, parent) end)
 
         {:ok, res}
@@ -36,9 +39,10 @@ defmodule TdDd.Lineage.GraphData.Nodes do
     end
   end
 
-  defp children(%Graph{} = t, id, roots) do
+  defp children(%Graph{} = t, id, roots, user) do
     t
     |> do_children(id, roots)
+    |> Enum.reject(&by_permissions(user, t, &1))
     |> Enum.reject(&hidden?(t, &1))
     |> Enum.map(&Graph.vertex(t, &1))
     |> Enum.map(&node_label/1)
@@ -51,6 +55,58 @@ defmodule TdDd.Lineage.GraphData.Nodes do
 
   defp hidden?(%Graph{} = t, id) do
     Graph.vertex(t, id, "hidden") == true
+  end
+
+  defp by_permissions(user, t, node) do
+    item = Units.get_node(node, preload: [:structure])
+
+    case has_permissions?(user, item) do
+      true ->
+        [node]
+        |> Traversal.reachable(t)
+        |> Enum.map(&Graph.vertex(t, &1))
+        |> Enum.map(&Map.get(&1, :id))
+        |> List.delete(node)
+        |> reject?(user)
+
+      false ->
+        true
+    end
+  end
+
+  defp has_permissions?(user, %Node{structure: %DataStructure{} = data_structure}) do
+    import Canada, only: [can?: 2]
+    can?(user, view_data_structure(data_structure))
+  end
+
+  defp has_permissions?(_user, _node), do: true
+
+  defp reject?([], _user), do: false
+
+  defp reject?(ids, user) do
+    resources = fetch_nodes(ids, "Resource")
+
+    case reject_collection?(resources, user) do
+      false ->
+        groups = fetch_nodes(ids, "Group")
+        reject_collection?(groups, user)
+
+      _ ->
+        true
+    end
+  end
+
+  defp fetch_nodes(ids, type) do
+    Map.new()
+    |> Map.put(:external_id, ids)
+    |> Map.put(:type, type)
+    |> Units.list_nodes(preload: [:structure])
+  end
+
+  defp reject_collection?([], _user), do: false
+
+  defp reject_collection?(collection, user) do
+    not Enum.any?(collection, &has_permissions?(user, &1))
   end
 
   defp class(%{class: "Group"}), do: :groups