+++ lib/td_dd/lineage/units.ex
@@ -61,16 +61,34 @@ defmodule TdDd.Lineage.Units do
     |> Enum.at(0)
   end
 
-  def list_nodes(clauses) do
+  def get_node(external_id, options \\ []) do
+    Node
+    |> Repo.get_by(external_id: external_id)
+    |> Repo.preload(options[:preload] || [])
+  end
+
+  def list_nodes(clauses, options \\ []) do
     clauses
     |> Map.new()
     |> Map.put_new(:deleted, false)
     |> Enum.reduce(Node, fn
-      {:type, type}, q -> where(q, [n], n.type == ^type)
-      {:deleted, false}, q -> where(q, [n], is_nil(n.deleted_at))
-      {:deleted, _true}, q -> q
+      {:external_id, external_ids}, q when is_list(external_ids) ->
+        where(q, [ds], ds.external_id in ^external_ids)
+
+      {:external_id, external_id}, q ->
+        where(q, [ds], ds.external_id == ^external_id)
+
+      {:type, type}, q ->
+        where(q, [n], n.type == ^type)
+
+      {:deleted, false}, q ->
+        where(q, [n], is_nil(n.deleted_at))
+
+      {:deleted, _true}, q ->
+        q
     end)
     |> Repo.all()
+    |> Repo.preload(options[:preload] || [])
   end
 
   def list_relations(clauses) do