+++ test/td_dd_web/controllers/node_controller_test.exs
@@ -2,6 +2,7 @@ defmodule TdDdWeb.NodeControllerTest do
   use TdDdWeb.ConnCase
   use TdDd.GraphDataCase
 
+  alias TdCache.TaxonomyCache
   alias TdDd.Lineage.GraphData
   alias TdDd.Lineage.GraphData.State
   alias TdDd.Permissions.MockPermissionResolver
@@ -10,13 +11,15 @@ defmodule TdDdWeb.NodeControllerTest do
   @admin_user_name "app-admin"
 
   setup_all do
+    stop_supervised(GraphData)
+    start_supervised(GraphData)
     start_supervised(MockTdAuthService)
     start_supervised(MockPermissionResolver)
     :ok
   end
 
   setup %{conn: conn} = tags do
-    start_supervised({GraphData, state: setup_state(tags)})
+    GraphData.state(state: setup_state(tags))
     {:ok, conn: put_req_header(conn, "accept", "application/json")}
   end
 
@@ -26,7 +29,9 @@ defmodule TdDdWeb.NodeControllerTest do
     @tag depends: [{"bar", "baz"}]
     test "index returns the top-level groups and parent nil", %{conn: conn} do
       conn = get(conn, Routes.node_path(conn, :index))
+
       assert [%{"parent" => nil, "groups" => [group]}] = json_response(conn, 200)["data"]
+
       assert %{"external_id" => "foo", "name" => "foo"} = group
     end
 
@@ -45,6 +50,186 @@ defmodule TdDdWeb.NodeControllerTest do
       assert %{"bar" => _bar, "baz" => _baz} = Enum.group_by(resources, & &1["external_id"])
       assert %{"bar" => _bar, "baz" => _baz} = Enum.group_by(resources, & &1["name"])
       assert %{"foo_type" => [_bar, _baz]} = Enum.group_by(resources, & &1["type"])
+    end
+
+    @tag authenticated_no_admin_user: "user"
+    @tag contains: %{"foo" => ["bar", "baz"], "xyz" => ["x", "y"]}
+    @tag depends: [{"bar", "baz"}, {"x", "y"}]
+    test "will filter groups depending on user permissions over domains", %{
+      conn: conn,
+      user: %{id: user_id}
+    } do
+      domain_id = :random.uniform(1_000_000)
+      TaxonomyCache.put_domain(%{name: "domain", id: domain_id, updated_at: DateTime.utc_now()})
+
+      MockPermissionResolver.create_acl_entry(%{
+        principal_id: user_id,
+        principal_type: "user",
+        resource_id: domain_id,
+        resource_type: "domain",
+        role_name: "no_perms"
+      })
+
+      insert(
+        :node,
+        external_id: "foo"
+      )
+
+      insert(
+        :node,
+        external_id: "xyz"
+      )
+
+      data_structure =
+        insert(
+          :data_structure,
+          external_id: "bar",
+          domain_id: domain_id
+        )
+
+      insert(:data_structure_version,
+        data_structure_id: data_structure.id,
+        name: data_structure.external_id
+      )
+
+      insert(
+        :node,
+        external_id: data_structure.external_id,
+        structure: data_structure
+      )
+
+      data_structure =
+        insert(
+          :data_structure,
+          external_id: "baz",
+          domain_id: domain_id
+        )
+
+      insert(:data_structure_version,
+        data_structure_id: data_structure.id,
+        name: data_structure.external_id
+      )
+
+      insert(
+        :node,
+        external_id: data_structure.external_id,
+        structure: data_structure
+      )
+
+      domain_id = :random.uniform(1_000_000)
+      TaxonomyCache.put_domain(%{name: "domain1", id: domain_id, updated_at: DateTime.utc_now()})
+
+      MockPermissionResolver.create_acl_entry(%{
+        principal_id: user_id,
+        principal_type: "user",
+        resource_id: domain_id,
+        resource_type: "domain",
+        role_name: "watch"
+      })
+
+      data_structure =
+        insert(
+          :data_structure,
+          external_id: "x",
+          domain_id: domain_id
+        )
+
+      insert(:data_structure_version,
+        data_structure_id: data_structure.id,
+        name: data_structure.external_id
+      )
+
+      insert(
+        :node,
+        external_id: data_structure.external_id,
+        structure: data_structure
+      )
+
+      data_structure =
+        insert(
+          :data_structure,
+          external_id: "y",
+          domain_id: domain_id
+        )
+
+      insert(:data_structure_version,
+        data_structure_id: data_structure.id,
+        name: data_structure.external_id
+      )
+
+      insert(
+        :node,
+        external_id: data_structure.external_id,
+        structure: data_structure
+      )
+
+      conn = get(conn, Routes.node_path(conn, :index))
+      assert [%{"parent" => nil, "groups" => [group]}] = json_response(conn, 200)["data"]
+
+      assert %{"external_id" => "xyz", "name" => "xyz"} = group
+    end
+
+    @tag authenticated_no_admin_user: "user"
+    @tag contains: %{"foo" => ["bar", "baz"]}
+    @tag depends: [{"bar", "baz"}]
+    test "will filter a group if a user has not permissions over any of the group's permissions", %{
+      conn: conn,
+      user: %{id: user_id}
+    } do
+      domain_id = :random.uniform(1_000_000)
+      TaxonomyCache.put_domain(%{name: "domain", id: domain_id, updated_at: DateTime.utc_now()})
+
+      MockPermissionResolver.create_acl_entry(%{
+        principal_id: user_id,
+        principal_type: "user",
+        resource_id: domain_id,
+        resource_type: "domain",
+        role_name: "no_perms"
+      })
+
+      insert(
+        :node,
+        external_id: "foo"
+      )
+
+      data_structure =
+        insert(
+          :data_structure,
+          external_id: "bar",
+          domain_id: domain_id
+        )
+
+      insert(:data_structure_version,
+        data_structure_id: data_structure.id,
+        name: data_structure.external_id
+      )
+
+      insert(
+        :node,
+        external_id: data_structure.external_id,
+        structure: data_structure
+      )
+
+      data_structure =
+        insert(
+          :data_structure,
+          external_id: "baz",
+          domain_id: domain_id
+        )
+
+      insert(:data_structure_version,
+        data_structure_id: data_structure.id,
+        name: data_structure.external_id
+      )
+
+      insert(
+        :node,
+        external_id: data_structure.external_id,
+        structure: data_structure
+      )
+
+      conn = get(conn, Routes.node_path(conn, :index))
+      assert [%{"parent" => nil}] = json_response(conn, 200)["data"]
     end
   end
 end